requirements.txt
streamlit
weaviate-client
sentence-transformers
unstructured
pdfminer.six
python-magic
pypandoc
unstructured[local-inference]
